pyproject.toml
[project]
name = "radical"
version = "0.1.0"
description = "The RAdio DIreactional Calibrator"
authors = [
    {name = "Torrance Hodgson",email = "torrance@pravic.xyz"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-casacore (>=3.6.1,<4.0.0)",
    "numba (>=0.61.0,<0.62.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "astropy (>=7.0.1,<8.0.0)",
    "mwa-hyperbeam (>=0.10.1,<0.11.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "cupy-cuda12x (>=13.4.1,<14.0.0)",
]

[project.scripts]
radical = "radical.cli:run"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
